PF-25 move auth actions to actions/auth
